apps/hooks/requirements.txt
# requests==2.32.3
# # python-dotenv==1.0.0
# linkedin-api==2.3.1
# fake-useragent==1.1.3
# attrs==24.2.0
# Automat==24.8.1
# Babel==2.17.0
# bcrypt==3.2.2
# blinker==1.7.0
# certifi==2024.8.30
# chardet==5.2.0
# click==8.1.7
# colorama==0.4.6
# configobj==5.0.8
# constantly==15.1.0
# cryptography==41.0.7
# distro==1.8.0
# greenlet==3.2.3
# httplib2==0.20.4
# hyperlink==21.0.0
# idna==3.10
# incremental==21.3.0
# Jinja2==3.1.2
# jsonpatch==1.32
# jsonpointer==2.3
# jsonschema==4.20.0
# markdown-it-py==2.2.0
# MarkupSafe==2.1.3
# mdurl==0.1.2
# msgpack==1.0.6
# # netifaces==0.11.0
# oauthlib==3.2.2
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# Pygments==2.18.0
# PyHamcrest==2.0.4
# PyJWT==2.8.0
# pyOpenSSL==23.2.0
# pyparsing==3.1.1
# pyrsistent==0.20.0
# pyserial==3.5
# pytz==2023.3
# PyYAML==6.0.2
# # rich==13.7.0
# service-identity==21.1.0
# setuptools==68.2.2
# Twisted==22.10.0
# urllib3==1.26.18
# wheel==0.44.0
# zope.interface==7.2 
#
# # Core dependencies
# pandas==2.2.3  # Newer version, likely has Python 3.13 wheels
# numpy==2.1.2  # Slightly newer, ensures compatibility
#
# # Web scraping
# seleniumbase==4.31.2
# selenium==4.25.0
# html2text==2025.4.15
# python-dateutil==2.9.0.post0
#
# # Data processing
# # python-dateutil==2.8.2
#
# # CLI
# typer==0.12.5
# rich==13.8.1
#
# # Configuration and validation
# pydantic==2.9.2  # Newer version, aligns with pydantic-core
# pydantic-core==2.23.4  # Newer version, likely has wheels
# pydantic-settings==2.5.2  # Newer version for compatibility
#
# # Development and testing
# pytest==8.3.3
# pytest-mock==3.12.0
# pytest-cov==4.1.0
#
# # Code quality
# black==23.12.1
# flake8==6.1.0
# isort==5.13.2
# mypy==1.8.0
#
# # Type annotations
# types-requests==2.31.0.10
# types-selenium==3.141.9
#
# # Config file support 
# python-dotenv==1.0.1
#
# # Captcha
# 2captcha-python==1.2.2

# Core dependencies
requests==2.32.4
pandas==2.2.3
numpy==2.1.2
python-dotenv==1.0.1  # Unified version
html2text==2025.4.15  # Unified, corrected version
rich==14.0.0  # Unified version

# LinkedIn scraper
linkedin-api==2.2.0
# fake-useragent==1.1.3
# attrs==25.3.0
# Automat==24.8.1
# Babel==2.17.0
# bcrypt==3.2.2
# blinker==1.7.0
# certifi==2025.6.15
# chardet==5.2.0
# click==8.1.7
# configobj==5.0.8
# constantly==15.1.0
# cryptography==41.0.7
# distro==1.8.0
# greenlet==3.2.3
# httplib2==0.20.4
# hyperlink==21.0.0
# idna==3.10
# incremental==24.7.0
# Jinja2==3.1.6
# jsonpatch==1.32
# jsonpointer==2.3
# jsonschema==4.20.0
# markdown-it-py==3.0.0
# MarkupSafe==3.0.2
# mdurl==0.1.2
# msgpack==1.0.6
# oauthlib==3.2.2
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# PyHamcrest==2.0.4
# PyJWT==2.8.0
# pyOpenSSL==23.2.0
# pyparsing==3.1.1
# pyrsistent==0.20.0
# pyserial==3.5
# pytz==2023.3
# PyYAML==6.0.2
# service-identity==21.1.0
# setuptools==80.9.0
# Twisted==24.7.0  # Upgraded for Python 3.13 compatibility
# urllib3==1.26.20
# wheel==0.45.1
# zope.interface==7.2

# Indeed scraper
seleniumbase==4.39.5
selenium==4.33.0
python-dateutil==2.9.0.post0
typer==0.12.5
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-settings==2.5.2

# Development and testing
pytest==8.4.0
pytest-mock==3.12.0
pytest-cov==4.1.0
black==24.8.0  # Upgraded
flake8==7.1.1  # Upgraded
isort==5.13.2
mypy==1.11.2  # Upgraded
types-requests==2.31.0.10
types-selenium==3.141.9
